fix(iac): reject a reused GCP WIF provider whose audience list excludes ours
An existing provider was compared on issuer, condition and mapping but not on
oidc.allowedAudiences. That list restricts which `aud` the provider accepts,
so one that omits this pool's audience rejects every token CUDly mints while
the script reports success over it. Same class as the soft-deleted and
disabled cases already handled here.

Empty is accepted rather than treated as unrestricted: GCP defaults an unset
list to the provider's own resource name, which is exactly what WIF_AUDIENCE
is. A non-empty list must contain it. gcloud's value() printer joins a
repeated field with ';', verified against the bundled SDK's printer rather
than assumed.

Also in this round:

- Corrects the charset comment on CUDLY_FEDERATED_SUBJECT and the test case
  that read as an endorsement of it. The comment justified '|' with
  Auth0/Okta-style subjects, a caller that cannot exist here, and the accept
  case asserted that google-oauth2|1234567890 "configures the project end to
  end". gcpFederatedSubject in internal/credentials/gcp_federated.go is a
  compile-time constant, so CUDly always presents 'cudly-controller' and any
  override pins the provider to a subject it will never sign. The case now
  asserts what it actually demonstrates, that the value round-trips through
  the template unmangled, and the character is documented as being in the set
  for parity with validate_principal_value in the ARM script.
- Adds the missing test for --oidc-credential-source-field without a source,
  which the ARM script rejects but nothing covered.
- Moves the GCP WIF stub and tests to templates_gcp_wif_test.go. The project
  caps files at 500 lines and templates_test.go had reached 1122; it is now
  577 and the new file 557. Kept in one commit with the changes above because
  the moved and original definitions collide in any intermediate state.

Refs #1661
